+++ packages/canary-react/src/component-tests/IcDataTable/IcDataTable.cy.tsx
@@ -90,6 +90,7 @@ const PAGINATION_GO_TO_PAGE_TEXT_FIELD_SELECTOR =
 const PAGINATION_GO_TO_PAGE_BUTTON_SELECTOR = ".go-to-page-holder ic-button";
 const ITEMS_PER_PAGE_SELECTOR = ".items-per-page-input";
 const ACTION_ELEMENT = "action-element";
+const TABLE_ROW_SELECTED = "table-row-selected";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/explicit-module-boundary-types
 export const BasicDataTable = (dataTableProps?: any): ReactElement => (
@@ -123,6 +124,40 @@ export const BasicSectionContainer = (): ReactElement => {
         </IcDataTable>
       </IcSectionContainer>
     </div>
+  );
+};
+
+export const ExternalSortDataTable = (): ReactElement => {
+  const ExternalData = [...DATA];
+  const handleSort = (detail: any) => {
+    if (detail.sorted === "ascending") {
+      ExternalData.sort((a, b) => {
+        const column = detail.columnName as keyof (typeof ExternalData)[0];
+        if (a[column] < b[column]) return -1;
+        if (a[column] > b[column]) return 1;
+        return 0;
+      });
+    } else if (detail.sorted === "descending") {
+      ExternalData.sort((a, b) => {
+        const column = detail.columnName as keyof (typeof ExternalData)[0];
+        if (a[column] < b[column]) return 1;
+        if (a[column] > b[column]) return -1;
+        return 0;
+      });
+    }
+  };
+  return (
+    <IcDataTable
+      columns={COLS}
+      data={ExternalData}
+      caption="Data Tables"
+      sortable
+      disableAutoSort
+      onIcSortChange={(e) => handleSort(e.detail)}
+      sortOptions={{
+        sortOrders: ["ascending", "descending"],
+      }}
+    />
   );
 };
 
@@ -369,6 +404,42 @@ describe("IcDataTables", () => {
       .find("td")
       .eq(1)
       .should(HAVE_TEXT, "Bartholomew Christoper Augustine Zacchaeus Ashford"); // cspell:disable-line
+  });
+
+  it("should sort data externally if disableAutoSort is true", () => {
+    mount(<ExternalSortDataTable />);
+
+    cy.checkHydrated(DATA_TABLE_SELECTOR);
+
+    cy.findShadowEl(DATA_TABLE_SELECTOR, SORT_BUTTON_SELECTOR)
+      .eq(0)
+      .shadow()
+      .find(TOOLTIP_BUTTON_SELECTOR)
+      .should(HAVE_ATTR, ARIA_LABEL, "Sort ascending");
+
+    cy.findShadowEl(DATA_TABLE_SELECTOR, SORT_BUTTON_SELECTOR).eq(0).click();
+
+    cy.findShadowEl(DATA_TABLE_SELECTOR, "tr")
+      .eq(1)
+      .find("td")
+      .eq(0)
+      .should(HAVE_TEXT, "Joe");
+
+    cy.findShadowEl(DATA_TABLE_SELECTOR, SORT_BUTTON_SELECTOR)
+      .eq(0)
+      .shadow()
+      .find(TOOLTIP_BUTTON_SELECTOR)
+      .should(HAVE_ATTR, ARIA_LABEL, "Sort descending");
+
+    cy.findShadowEl(DATA_TABLE_SELECTOR, SORT_BUTTON_SELECTOR).eq(0).click();
+
+    cy.findShadowEl(DATA_TABLE_SELECTOR, "tr")
+      .eq(1)
+      .find("td")
+      .eq(0)
+      .should(HAVE_TEXT, "Sarah");
+
+    cy.checkA11yWithWait();
   });
 
   it("should render with column overrides", () => {
@@ -1091,7 +1162,7 @@ describe("IcDataTables", () => {
     cy.findShadowEl(DATA_TABLE_SELECTOR, "td")
       .eq(0)
       .find("span")
-      .should(HAVE_CLASS, "action-element")
+      .should(HAVE_CLASS, ACTION_ELEMENT)
       .find("ic-button")
       .should("be.visible");
 
@@ -1142,7 +1213,7 @@ describe("IcDataTables", () => {
       .should(HAVE_CSS, "grid-template-columns", "156.797px 32px");
 
     cy.findShadowEl(DATA_TABLE_SELECTOR, "span")
-      .should(HAVE_CLASS, "action-element")
+      .should(HAVE_CLASS, ACTION_ELEMENT)
       .should(HAVE_CSS, "justify-content", "right");
   });
 
@@ -4470,6 +4541,38 @@ describe("IcDataTable row selection", () => {
         capture: "viewport",
       },
     });
+  });
+
+  it("should not highlight the selected row when the action element click event is stopped from propagating", () => {
+    mount(
+      <IcDataTable
+        columns={COLS}
+        data={ACTION_DATA_ELEMENTS}
+        caption="Data tables"
+      ></IcDataTable>
+    );
+
+    cy.checkHydrated(DATA_TABLE_SELECTOR);
+
+    cy.get(DATA_TABLE_SELECTOR)
+      .shadow()
+      .find(`[data-testid="copy-button"]`)
+      .eq(0)
+      .click();
+
+    cy.findShadowEl(DATA_TABLE_SELECTOR, "tr")
+      .eq(1)
+      .should(NOT_HAVE_CLASS, TABLE_ROW_SELECTED);
+
+    cy.get(DATA_TABLE_SELECTOR)
+      .shadow()
+      .find(`[data-testid="copy-button"]`)
+      .eq(1)
+      .click();
+
+    cy.findShadowEl(DATA_TABLE_SELECTOR, "tr")
+      .eq(2)
+      .should(HAVE_CLASS, TABLE_ROW_SELECTED);
   });
 
   it("should emit icSelectedRowChange event when the selected row changes", () => {