+++ packages/canary-web-components/src/components/ic-data-table/ic-data-table.tsx
@@ -171,6 +171,11 @@ export class DataTable {
   @Prop() data?: IcDataTableDataType[];
 
   /**
+   * If `true`, the built in sort functionality will be disabled. For example, if rows will already be sorted from an external source.
+   */
+  @Prop() disableAutoSort?: boolean = false;
+
+  /**
    * Set the density of the table including font and padding.
    */
   @Prop({ mutable: true }) density?: IcDataTableDensityOptions = "default";
@@ -1525,7 +1530,7 @@ export class DataTable {
                   // eslint-disable-next-line react/jsx-no-bind
                   onClick={
                     cell.actionOnClick
-                      ? () => this.handleClick(cell.actionOnClick)
+                      ? (event) => this.handleClick(event, cell.actionOnClick)
                       : undefined
                   }
                 ></span>
@@ -1679,7 +1684,11 @@ export class DataTable {
     });
 
     return organisedData
-      .sort(!this.sortable ? undefined : this.getSortFunction())
+      .sort(
+        !this.sortable || this.disableAutoSort
+          ? undefined
+          : this.getSortFunction()
+      )
       .map((row, index) => {
         return (
           <tr
@@ -2113,7 +2122,8 @@ export class DataTable {
     });
   };
 
-  private handleClick = (callback: () => void) => callback();
+  private handleClick = (event: Event, callback: (event: Event) => void) =>
+    callback(event);
 
   private renderTableBody = (
     data: IcDataTableDataType[],