+++ packages/web-components/src/components/ic-badge/ic-badge.tsx
@@ -6,6 +6,7 @@ import {
   Prop,
   Watch,
   h,
+  Listen,
 } from "@stencil/core";
 import {
   IcBadgePositions,
@@ -27,6 +28,9 @@ import {
   onComponentRequiredPropUndefined,
   rgbaStrToObj,
 } from "../../utils/helpers";
+
+const NAVIGATION_BUTTON = "IC-NAVIGATION-BUTTON";
+const TOP_NAVIGATION = "IC-TOP-NAVIGATION";
 
 /**
  * @slot badge-icon - Icon will be rendered inside the badge if type is set to icon.
@@ -76,7 +80,7 @@ export class Badge {
   /**
    * The positioning of the badge in reference to the parent element.
    */
-  @Prop() position?: IcBadgePositions = "far";
+  @Prop({ mutable: true }) position?: IcBadgePositions = "far";
 
   /**
    * The size of the badge to be displayed.
@@ -129,6 +133,20 @@ export class Badge {
         [{ prop: this.textLabel, propName: "text-label" }],
         "Badge"
       );
+  }
+
+  componentWillRender(): void {
+    this.isInTopNav() && this.setPositionInTopNavigation();
+  }
+
+  @Listen("icNavigationMenuOpened", { target: "document" })
+  navBarMenuOpenHandler(): void {
+    this.isInTopNav() && (this.position = "inline");
+  }
+
+  @Listen("icNavigationMenuClosed", { target: "document" })
+  navBarMenuCloseHandler(): void {
+    this.isInTopNav() && (this.position = "near");
   }
 
   /**
@@ -213,6 +231,22 @@ export class Badge {
         this.ariaLabel = `, ${defaultAriaLabel}`;
       }
     }
+  };
+
+  private setPositionInTopNavigation = () => {
+    const parentTopNavEl = this.el.parentElement.parentElement;
+    parentTopNavEl.classList.contains("mobile-mode")
+      ? (this.position = "inline")
+      : (this.position = "near");
+  };
+
+  private isInTopNav = (): boolean => {
+    const parentEl = this.el.parentElement;
+    const grandparentEl = parentEl.parentElement;
+    return (
+      parentEl.tagName === NAVIGATION_BUTTON &&
+      grandparentEl.tagName === TOP_NAVIGATION
+    );
   };
 
   private isAccessibleLabelDefined = () =>