+++ packages/fonts/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.6.6 (2025-02-19)
+
+**Note:** Version bump only for package @ukic/fonts
+
 ## 2.6.5 (2024-12-18)
 
 ### Bug Fixes